Added awkward ratings for android
Simple app that uses the REST APIs documented on www.themoviedb.org. Specifically, it asks for the day's most popular movies (sorted by when that movie was last updated in my database) and allows users of the app to vote on which movies would be awkward to watch with parents (or vice versa). Results are presented in an endless RecyclerView.

Used open source libraries:

Retrofit
Picasso
AppCompat-v7's Palette